+++ src/main/java/com/gxdxx/instagram/domain/comment/dao/CommentRepository.java
@@ -1,4 +1,4 @@
-package com.gxdxx.instagram.repository;
+package com.gxdxx.instagram.domain.comment.dao;
 
 import com.gxdxx.instagram.domain.comment.domain.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
