+++ src/main/java/com/gxdxx/instagram/service/chatroom/ChatRoomQueryService.java
@@ -5,8 +5,8 @@ import com.gxdxx.instagram.dto.response.ChatRoomListResponse;
 import com.gxdxx.instagram.dto.response.ChatRoomResponse;
 import com.gxdxx.instagram.domain.user.domain.User;
 import com.gxdxx.instagram.exception.UserNotFoundException;
-import com.gxdxx.instagram.repository.ChatRoomRepository;
-import com.gxdxx.instagram.repository.UserRepository;
+import com.gxdxx.instagram.domain.chatroom.dao.ChatRoomRepository;
+import com.gxdxx.instagram.domain.user.dao.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
