+++ src/main/java/com/gxdxx/instagram/service/comment/CommentCreateService.java
@@ -7,9 +7,9 @@ import com.gxdxx.instagram.domain.post.domain.Post;
 import com.gxdxx.instagram.domain.user.domain.User;
 import com.gxdxx.instagram.exception.PostNotFoundException;
 import com.gxdxx.instagram.exception.UserNotFoundException;
-import com.gxdxx.instagram.repository.CommentRepository;
-import com.gxdxx.instagram.repository.PostRepository;
-import com.gxdxx.instagram.repository.UserRepository;
+import com.gxdxx.instagram.domain.comment.dao.CommentRepository;
+import com.gxdxx.instagram.domain.post.dao.PostRepository;
+import com.gxdxx.instagram.domain.user.dao.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
