+++ src/main/java/com/gxdxx/instagram/service/follow/FollowDeleteService.java
@@ -5,8 +5,8 @@ import com.gxdxx.instagram.domain.follow.domain.Follow;
 import com.gxdxx.instagram.domain.user.domain.User;
 import com.gxdxx.instagram.exception.FollowNotFountException;
 import com.gxdxx.instagram.exception.UserNotFoundException;
-import com.gxdxx.instagram.repository.FollowRepository;
-import com.gxdxx.instagram.repository.UserRepository;
+import com.gxdxx.instagram.domain.follow.dao.FollowRepository;
+import com.gxdxx.instagram.domain.user.dao.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
