+++ src/main/java/com/gxdxx/instagram/service/message/MessageCreateService.java
@@ -7,9 +7,9 @@ import com.gxdxx.instagram.domain.message.domain.Message;
 import com.gxdxx.instagram.domain.user.domain.User;
 import com.gxdxx.instagram.exception.InvalidRequestException;
 import com.gxdxx.instagram.exception.UserNotFoundException;
-import com.gxdxx.instagram.repository.ChatRoomRepository;
-import com.gxdxx.instagram.repository.MessageRepository;
-import com.gxdxx.instagram.repository.UserRepository;
+import com.gxdxx.instagram.domain.chatroom.dao.ChatRoomRepository;
+import com.gxdxx.instagram.domain.message.dao.MessageRepository;
+import com.gxdxx.instagram.domain.user.dao.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
