+++ src/main/java/com/gxdxx/instagram/service/message/MessageQueryService.java
@@ -8,9 +8,9 @@ import com.gxdxx.instagram.domain.user.domain.User;
 import com.gxdxx.instagram.exception.ChatRoomNotFoundException;
 import com.gxdxx.instagram.exception.UnauthorizedAccessException;
 import com.gxdxx.instagram.exception.UserNotFoundException;
-import com.gxdxx.instagram.repository.ChatRoomRepository;
-import com.gxdxx.instagram.repository.MessageRepository;
-import com.gxdxx.instagram.repository.UserRepository;
+import com.gxdxx.instagram.domain.chatroom.dao.ChatRoomRepository;
+import com.gxdxx.instagram.domain.message.dao.MessageRepository;
+import com.gxdxx.instagram.domain.user.dao.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
