+++ src/main/java/com/gxdxx/instagram/service/post/PostFeedQueryService.java
@@ -5,8 +5,8 @@ import com.gxdxx.instagram.dto.response.FeedResponse;
 import com.gxdxx.instagram.dto.response.PostFeedResponse;
 import com.gxdxx.instagram.domain.user.domain.User;
 import com.gxdxx.instagram.exception.UserNotFoundException;
-import com.gxdxx.instagram.repository.PostRepository;
-import com.gxdxx.instagram.repository.UserRepository;
+import com.gxdxx.instagram.domain.post.dao.PostRepository;
+import com.gxdxx.instagram.domain.user.dao.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
