+++ src/main/java/com/gxdxx/instagram/service/reply/ReplyCreateService.java
@@ -7,9 +7,9 @@ import com.gxdxx.instagram.domain.reply.domain.Reply;
 import com.gxdxx.instagram.domain.user.domain.User;
 import com.gxdxx.instagram.exception.CommentNotFoundException;
 import com.gxdxx.instagram.exception.UserNotFoundException;
-import com.gxdxx.instagram.repository.CommentRepository;
-import com.gxdxx.instagram.repository.ReplyRepository;
-import com.gxdxx.instagram.repository.UserRepository;
+import com.gxdxx.instagram.domain.comment.dao.CommentRepository;
+import com.gxdxx.instagram.domain.reply.dao.ReplyRepository;
+import com.gxdxx.instagram.domain.user.dao.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
