+++ src/main/java/com/gxdxx/instagram/service/reply/ReplyUpdateService.java
@@ -5,7 +5,7 @@ import com.gxdxx.instagram.dto.response.ReplyUpdateResponse;
 import com.gxdxx.instagram.domain.reply.domain.Reply;
 import com.gxdxx.instagram.exception.ReplyNotFoundException;
 import com.gxdxx.instagram.exception.UnauthorizedAccessException;
-import com.gxdxx.instagram.repository.ReplyRepository;
+import com.gxdxx.instagram.domain.reply.dao.ReplyRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
