+++ src/main/java/com/gxdxx/instagram/service/user/UserProfileQueryService.java
@@ -3,8 +3,8 @@ package com.gxdxx.instagram.service.user;
 import com.gxdxx.instagram.dto.response.UserProfileResponse;
 import com.gxdxx.instagram.domain.user.domain.User;
 import com.gxdxx.instagram.exception.UserNotFoundException;
-import com.gxdxx.instagram.repository.FollowRepository;
-import com.gxdxx.instagram.repository.UserRepository;
+import com.gxdxx.instagram.domain.follow.dao.FollowRepository;
+import com.gxdxx.instagram.domain.user.dao.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
