+++ src/test/java/com/gxdxx/instagram/service/follow/FollowCreateServiceTest.java
@@ -7,8 +7,8 @@ import com.gxdxx.instagram.domain.user.domain.User;
 import com.gxdxx.instagram.exception.FollowAlreadyExistsException;
 import com.gxdxx.instagram.exception.InvalidRequestException;
 import com.gxdxx.instagram.exception.UserNotFoundException;
-import com.gxdxx.instagram.repository.FollowRepository;
-import com.gxdxx.instagram.repository.UserRepository;
+import com.gxdxx.instagram.domain.follow.dao.FollowRepository;
+import com.gxdxx.instagram.domain.user.dao.UserRepository;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
