+++ src/test/java/com/gxdxx/instagram/service/post/PostCreateServiceTest.java
@@ -5,8 +5,8 @@ import com.gxdxx.instagram.dto.response.PostRegisterResponse;
 import com.gxdxx.instagram.domain.post.domain.Post;
 import com.gxdxx.instagram.domain.user.domain.User;
 import com.gxdxx.instagram.exception.UserNotFoundException;
-import com.gxdxx.instagram.repository.PostRepository;
-import com.gxdxx.instagram.repository.UserRepository;
+import com.gxdxx.instagram.domain.post.dao.PostRepository;
+import com.gxdxx.instagram.domain.user.dao.UserRepository;
 import com.gxdxx.instagram.config.s3.S3Uploader;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
