+++ src/test/java/com/gxdxx/instagram/service/user/UserProfileQueryServiceTest.java
@@ -3,8 +3,8 @@ package com.gxdxx.instagram.service.user;
 import com.gxdxx.instagram.dto.response.UserProfileResponse;
 import com.gxdxx.instagram.domain.user.domain.User;
 import com.gxdxx.instagram.exception.UserNotFoundException;
-import com.gxdxx.instagram.repository.FollowRepository;
-import com.gxdxx.instagram.repository.UserRepository;
+import com.gxdxx.instagram.domain.follow.dao.FollowRepository;
+import com.gxdxx.instagram.domain.user.dao.UserRepository;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
